backend/requirements.txt
fastapi
uvicorn
playwright
pydantic
python-dotenv
httpx
